requirements.txt
flask
fastapi
uvicorn
biopython
python-multipart